Create README.md
A Students Information System is a centralized platform designed to manage
and streamline the administrative and academic functions of an educational
institution. The system handles various aspects such as student enrolment,
academic performance, attendance, and personal information. The main
objective of this system is to provide an efficient way to store, retrieve, and
manage student data, improving the overall educational experience for students,
teachers, and administrators.